# Use bbox col to speed up tile query exec time